requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-huggingface
sentence-transformers
chromadb
openai
akshare
pandas
apscheduler
crewai
crewai-tools
mem0ai
pydantic